Fix nightly-sdk-generation (#1983)
* Fix make clean in case 'bin' dir doesn't exist yet #1982

* Cleanup step should remove the whole SDKs

* Fetch git tags for pulumictl #1982

* Fetch git tags via fetch-depth instead #1982

+++ Makefile
@@ -95,7 +95,7 @@ lint_provider: provider/.mod_download.sentinel provider/cmd/$(PROVIDER)/*.go $(P
 	cd provider && GOGC=20 golangci-lint run -c ../.golangci.yml
 
 clean:
-	find bin -maxdepth 1 -type f -delete
+	if [ -d bin ]; then find bin -maxdepth 1 -type f -delete; fi
 	rm -rf nuget
 	find . -maxdepth 2 -name "*.sentinel" -delete
 	cd provider/cmd/arm2pulumi && rm -f metadata-compact.json schema-full.json
